+++ common/src/main/java/com/gregtechceu/gtceu/api/capability/recipe/IO.java
@@ -1,13 +1,24 @@
 package com.gregtechceu.gtceu.api.capability.recipe;
 
+import com.lowdragmc.lowdraglib.gui.texture.IGuiTexture;
+import com.lowdragmc.lowdraglib.gui.texture.ResourceTexture;
+
 /**
  * The capability can be input or output or both
  */
 public enum IO {
-    IN,
-    OUT,
-    BOTH,
-    NONE;
+    IN("gtceu.io.import", "import"),
+    OUT("gtceu.io.export", "export"),
+    BOTH("gtceu.io.both", "both"),
+    NONE("gtceu.io.none", "none");
+
+    public final String localeName;
+    public final IGuiTexture icon;
+
+    IO(String localeName, String textureName) {
+        this.localeName = localeName;
+        this.icon = new ResourceTexture("gtceu:textures/gui/icon/io_mode/" + textureName + ".png");
+    }
 
     public boolean support(IO io) {
         if (io == this) return true;